refactor: renaming of environment variables
