+++ client/src/layouts/Inside.astro
@@ -41,7 +41,7 @@ const { title } = Astro.props;
 		var HEIGHT = window.innerHeight;
 		var WAVEHEIGHT = 60;
 		var FREQUENCY = 200;
-		var SPEED = 2;
+		var SPEED = 1;
 
 		let xs: any = [];
 		let tick = 0;
