+++ client/src/layouts/Outside.astro
@@ -42,7 +42,7 @@ const { title } = Astro.props;
 		var HEIGHT = window.innerHeight;
 		var WAVEHEIGHT = 60;
 		var FREQUENCY = 200;
-		var SPEED = 2;
+		var SPEED = 1;
 
 		let xs: any = [];
 		let tick = 0;
