+++ client/src/services/device.ts
@@ -4,7 +4,7 @@ import { ErrorHandler } from "../lib/errors";
  * @class Handler of all OLT device requests for the API.
  */
 export class DeviceService {
-    private static url: string = process.env.PUBLIC_API_CORE ?? import.meta.env.PUBLIC_API_CORE;
+    private static url: string = import.meta.env.PUBLIC_API_CORE;
 
 
     /**
@@ -12,6 +12,7 @@ export class DeviceService {
      */
     static async getDevices(): Promise<{ status: (number | null), info: (any | ErrorHandler) }> {
         try {
+            console.log(this.url);
             const response = await fetch(`${this.url}/info/device`, {
                 method: 'GET'
             });
