+++ client/src/services/odn.ts
@@ -4,7 +4,7 @@ import { ErrorHandler } from "../lib/errors";
  * @class Handler of all ODN requests for the API.
  */
 export class OdnService {
-    private static url: string = process.env.PUBLIC_API_CORE ?? import.meta.env.PUBLIC_API_CORE;
+    private static url: string = import.meta.env.PUBLIC_API_CORE;
 
     /**
      * Request API to get all ODN by an OLT device ID.
